[front] refactor(agents): move batch delete to lib/api/assistant/configuration/agent.ts
Group batchHardDeletePendingAgentConfigurations alongside
unsafeHardDeleteAgentConfiguration per review feedback.
